+++ contracts/cw1-subkeys/src/whitelist.rs
@@ -1,8 +1,8 @@
 use cosmwasm_std::{Deps, DepsMut, Env, MessageInfo, Response, StdResult};
 use cw1_whitelist::responses::AdminListResponse;
-use cw1_whitelist::whitelist::Whitelist;
 #[cfg(test)]
-use cw1_whitelist::whitelist::{ExecMsg, QueryMsg};
+use cw1_whitelist::whitelist;
+use cw1_whitelist::whitelist::Whitelist;
 use sylvia::contract;
 
 use crate::contract::Cw1SubkeysContract;
@@ -11,6 +11,7 @@ use crate::error::ContractError;
 // This can be skipped by tarpaulin as it's covered in cw1-whitelist
 #[cfg(not(tarpaulin_include))]
 #[contract]
+#[messages(whitelist as Whitelist)]
 impl Whitelist for Cw1SubkeysContract<'_> {
     type Error = ContractError;