+++ contracts/cw1-whitelist/src/contract.rs
@@ -1,9 +1,5 @@
-#[cfg(test)]
-use crate::cw1::multitest_utils::Cw1Proxy;
 use crate::error::ContractError;
 use crate::whitelist;
-#[cfg(test)]
-use crate::whitelist::multitest_utils::WhitelistProxy;
 use cosmwasm_std::{Addr, Deps, DepsMut, Empty, Env, MessageInfo, Response};
 
 use cw2::set_contract_version;