+++ sylvia-derive/src/multitest.rs
@@ -4,7 +4,7 @@ use proc_macro_error::emit_error;
 use quote::quote;
 use syn::parse::{Parse, Parser};
 use syn::spanned::Spanned;
-use syn::{FnArg, GenericParam, ImplItem, ItemImpl, Pat, PatType, Path, Type};
+use syn::{FnArg, GenericParam, ImplItem, ItemImpl, ItemTrait, Pat, PatType, Path, Type};
 
 use crate::check_generics::CheckGenerics;
 use crate::crate_module;
@@ -59,6 +59,7 @@ impl<'a> MultitestHelpers<'a> {
         generics: &'a [&'a GenericParam],
     ) -> Self {
         let mut is_migrate = false;
+        let sylvia = crate_module();
 
         let messages: Vec<_> = source
             .items
@@ -91,7 +92,7 @@ impl<'a> MultitestHelpers<'a> {
                             }
                         }
                     } else {
-                        quote! { cw_multi_test::AppResponse }
+                        quote! { #sylvia ::cw_multi_test::AppResponse }
                     };
 
                     let name = &sig.ident;
@@ -196,9 +197,14 @@ impl<'a> MultitestHelpers<'a> {
             error_type,
             proxy_name,
             source,
+            is_trait,
             ..
         } = self;
         let sylvia = crate_module();
+
+        if *is_trait {
+            return self.impl_trait_on_proxy();
+        }
 
         let messages = messages.iter().map(|msg| {
             let MessageSignature {
@@ -244,16 +250,16 @@ impl<'a> MultitestHelpers<'a> {
                     Ok(interface) => {
                         let ContractMessageAttr { module, .. } = interface;
                         assert!(!module.segments.is_empty());
-                        let module = &module.segments[0].ident;
-                        let method_name = Ident::new(&format!("{}_proxy", module), module.span());
+                        let module_name = &module.segments[0].ident;
+                        let method_name = Ident::new(&format!("{}_proxy", module_name), module_name.span());
                         let proxy_name = Ident::new(
-                            &format!("{}Proxy", module.to_string().to_case(Case::UpperCamel)),
-                            module.span(),
+                            &format!("{}Proxy", module_name.to_string().to_case(Case::UpperCamel)),
+                            module_name.span(),
                         );
 
                         quote! {
-                            pub fn #method_name (&self) -> #proxy_name <'app> {
-                                #proxy_name ::new(self.contract_addr.clone(), self.app)
+                            pub fn #method_name (&self) -> #module ::trait_utils:: #proxy_name <'app> {
+                                #module ::trait_utils:: #proxy_name ::new(self.contract_addr.clone(), self.app)
                             }
                         }
                     }
@@ -268,18 +274,17 @@ impl<'a> MultitestHelpers<'a> {
             .collect();
 
         quote! {
-            #[cfg(test)]
             pub mod multitest_utils {
                 use super::*;
-                use cw_multi_test::Executor;
+                use #sylvia ::cw_multi_test::Executor;
 
                 pub struct #proxy_name <'app> {
-                    pub contract_addr: cosmwasm_std::Addr,
+                    pub contract_addr: #sylvia ::cw_std::Addr,
                     pub app: &'app #sylvia ::multitest::App,
                 }
 
                 impl<'app> #proxy_name <'app> {
-                    pub fn new(contract_addr: cosmwasm_std::Addr, app: &'app #sylvia ::multitest::App) -> Self {
+                    pub fn new(contract_addr: #sylvia ::cw_std::Addr, app: &'app #sylvia ::multitest::App) -> Self {
                         #proxy_name{ contract_addr, app }
                     }
 
@@ -288,13 +293,130 @@ impl<'a> MultitestHelpers<'a> {
                     #(#interfaces)*
                 }
 
-                impl<'app> From<(cosmwasm_std::Addr, &'app #sylvia ::multitest::App)> for #proxy_name<'app> {
-                    fn from(input: (cosmwasm_std::Addr, &'app #sylvia ::multitest::App)) -> #proxy_name<'app> {
+                impl<'app> From<(#sylvia ::cw_std::Addr, &'app #sylvia ::multitest::App)> for #proxy_name<'app> {
+                    fn from(input: (#sylvia ::cw_std::Addr, &'app #sylvia ::multitest::App)) -> #proxy_name<'app> {
                         #proxy_name::new(input.0, input.1)
                     }
                 }
 
                 #contract_block
+            }
+        }
+    }
+
+    fn impl_trait_on_proxy(&self) -> TokenStream {
+        let Self {
+            messages,
+            error_type,
+            source,
+            ..
+        } = self;
+
+        let sylvia = crate_module();
+
+        let interface_name = interface_name(self.source);
+        let proxy_name = &self.proxy_name;
+        let trait_name = Ident::new(&format!("{}Methods", interface_name), interface_name.span());
+
+        let modules: Vec<_> = source
+            .attrs
+            .iter()
+            .filter(|attr| attr.path.is_ident("messages"))
+            .filter_map(
+                |attr| match ContractMessageAttr::parse.parse2(attr.tokens.clone()) {
+                    Ok(interface) => {
+                        let ContractMessageAttr { module, .. } = &interface;
+                        assert!(!module.segments.is_empty());
+                        Some(module.segments[0].ident.clone())
+                    }
+                    Err(err) => {
+                        emit_error!(attr.span(), err);
+                        None
+                    }
+                },
+            )
+            .collect();
+
+        let module = match modules.len() {
+            0 => {
+                quote! {}
+            }
+            1 => {
+                let module = &modules[0];
+                quote! {#module ::}
+            }
+            _ => {
+                emit_error!(
+                    source.span(),
+                    "Only one #[messages] attribute is allowed per contract"
+                );
+                return quote! {};
+            }
+        };
+        let methods_definitions = messages.iter().map(|msg| {
+            let MessageSignature {
+                name,
+                params,
+                arguments,
+                msg_ty,
+                return_type,
+            } = msg;
+            if msg_ty == &MsgType::Exec {
+                quote! {
+                    #[track_caller]
+                    fn #name (&self, #(#params,)* ) -> #sylvia ::multitest::ExecProxy::<#error_type, #module ExecMsg> {
+                        let msg = #module ExecMsg:: #name ( #(#arguments),* );
+
+                        #sylvia ::multitest::ExecProxy::new(&self.contract_addr, msg, &self.app)
+                    }
+                }
+            } else {
+                quote! {
+                    fn #name (&self, #(#params,)* ) -> Result<#return_type, #error_type> {
+                        let msg = #module QueryMsg:: #name ( #(#arguments),* );
+
+                        self.app
+                            .app
+                            .borrow()
+                            .wrap()
+                            .query_wasm_smart(self.contract_addr.clone(), &msg)
+                            .map_err(Into::into)
+                    }
+                }
+            }
+        });
+
+        let methods_declarations = messages.iter().map(|msg| {
+            let MessageSignature {
+                name,
+                params,
+                msg_ty,
+                return_type,
+                ..
+            } = msg;
+            if msg_ty == &MsgType::Exec {
+                quote! {
+                    fn #name (&self, #(#params,)* ) -> #sylvia ::multitest::ExecProxy::<#error_type, #module ExecMsg>;
+                }
+            } else {
+                quote! {
+                    fn #name (&self, #(#params,)* ) -> Result<#return_type, #error_type>;
+                }
+            }
+        });
+
+        quote! {
+            #[cfg(test)]
+            pub mod test_utils {
+                use super::*;
+
+                pub trait #trait_name {
+                    #(#methods_declarations)*
+                }
+
+                impl #trait_name for #module trait_utils:: #proxy_name<'_> {
+                    #(#methods_definitions)*
+                }
             }
         }
     }
@@ -368,14 +490,14 @@ impl<'a> MultitestHelpers<'a> {
 
             pub struct InstantiateProxy<'a, 'app> {
                 code_id: &'a #code_id <'app>,
-                funds: &'a [cosmwasm_std::Coin],
+                funds: &'a [#sylvia ::cw_std::Coin],
                 label: &'a str,
                 admin: Option<String>,
                 msg: InstantiateMsg,
             }
 
             impl<'a, 'app> InstantiateProxy<'a, 'app> {
-                pub fn with_funds(self, funds: &'a [cosmwasm_std::Coin]) -> Self {
+                pub fn with_funds(self, funds: &'a [#sylvia ::cw_std::Coin]) -> Self {
                     Self { funds, ..self }
                 }
 
@@ -396,7 +518,7 @@ impl<'a> MultitestHelpers<'a> {
                         .borrow_mut()
                         .instantiate_contract(
                             self.code_id.code_id,
-                            cosmwasm_std::Addr::unchecked(sender),
+                            #sylvia ::cw_std::Addr::unchecked(sender),
                             &self.msg,
                             self.funds,
                             self.label,
@@ -414,81 +536,119 @@ impl<'a> MultitestHelpers<'a> {
 
     fn generate_impl_contract(&self) -> TokenStream {
         let contract = &self.contract;
+        let sylvia = crate_module();
 
         // MigrateMsg is not generated all the time in contrary to Exec, Query and Instantiate.
         let migrate_body = if self.is_migrate {
             quote! {
-                cosmwasm_std::from_slice::<MigrateMsg>(&msg)?
+                #sylvia ::cw_std::from_slice::<MigrateMsg>(&msg)?
                     .dispatch(self, (deps, env))
                     .map_err(Into::into)
             }
         } else {
             quote! {
-                anyhow::bail!("migrate not implemented for contract")
+                #sylvia ::anyhow::bail!("migrate not implemented for contract")
             }
         };
         quote! {
-            impl cw_multi_test::Contract<cosmwasm_std::Empty> for #contract {
+            impl #sylvia ::cw_multi_test::Contract<#sylvia ::cw_std::Empty> for #contract {
                 fn execute(
                     &self,
-                    deps: cosmwasm_std::DepsMut<cosmwasm_std::Empty>,
-                    env: cosmwasm_std::Env,
-                    info: cosmwasm_std::MessageInfo,
+                    deps: #sylvia ::cw_std::DepsMut<#sylvia ::cw_std::Empty>,
+                    env: #sylvia ::cw_std::Env,
+                    info: #sylvia ::cw_std::MessageInfo,
                     msg: Vec<u8>,
-                ) -> anyhow::Result<cosmwasm_std::Response<cosmwasm_std::Empty>> {
-                    cosmwasm_std::from_slice::<ContractExecMsg>(&msg)?
+                ) -> #sylvia ::anyhow::Result<#sylvia ::cw_std::Response<#sylvia ::cw_std::Empty>> {
+                    #sylvia ::cw_std::from_slice::<ContractExecMsg>(&msg)?
                         .dispatch(self, (deps, env, info))
                         .map_err(Into::into)
                 }
 
                 fn instantiate(
                     &self,
-                    deps: cosmwasm_std::DepsMut<cosmwasm_std::Empty>,
-                    env: cosmwasm_std::Env,
-                    info: cosmwasm_std::MessageInfo,
+                    deps: #sylvia ::cw_std::DepsMut<#sylvia ::cw_std::Empty>,
+                    env: #sylvia ::cw_std::Env,
+                    info: #sylvia ::cw_std::MessageInfo,
                     msg: Vec<u8>,
-                ) -> anyhow::Result<cosmwasm_std::Response<cosmwasm_std::Empty>> {
-                    cosmwasm_std::from_slice::<InstantiateMsg>(&msg)?
+                ) -> #sylvia ::anyhow::Result<#sylvia ::cw_std::Response<#sylvia ::cw_std::Empty>> {
+                    #sylvia ::cw_std::from_slice::<InstantiateMsg>(&msg)?
                         .dispatch(self, (deps, env, info))
                         .map_err(Into::into)
                 }
 
                 fn query(
                     &self,
-                    deps: cosmwasm_std::Deps<cosmwasm_std::Empty>,
-                    env: cosmwasm_std::Env,
+                    deps: #sylvia ::cw_std::Deps<#sylvia ::cw_std::Empty>,
+                    env: #sylvia ::cw_std::Env,
                     msg: Vec<u8>,
-                ) -> anyhow::Result<cosmwasm_std::Binary> {
-                    cosmwasm_std::from_slice::<ContractQueryMsg>(&msg)?
+                ) -> #sylvia ::anyhow::Result<#sylvia ::cw_std::Binary> {
+                    #sylvia ::cw_std::from_slice::<ContractQueryMsg>(&msg)?
                         .dispatch(self, (deps, env))
                         .map_err(Into::into)
                 }
 
                 fn sudo(
                     &self,
-                    _deps: cosmwasm_std::DepsMut<cosmwasm_std::Empty>,
-                    _env: cosmwasm_std::Env,
+                    _deps: #sylvia ::cw_std::DepsMut<#sylvia ::cw_std::Empty>,
+                    _env: #sylvia ::cw_std::Env,
                     _msg: Vec<u8>,
-                ) -> anyhow::Result<cosmwasm_std::Response<cosmwasm_std::Empty>> {
-                    anyhow::bail!("sudo not implemented for contract")
+                ) -> #sylvia ::anyhow::Result<#sylvia ::cw_std::Response<#sylvia ::cw_std::Empty>> {
+                    #sylvia ::anyhow::bail!("sudo not implemented for contract")
                 }
 
                 fn reply(
                     &self,
-                    _deps: cosmwasm_std::DepsMut<cosmwasm_std::Empty>,
-                    _env: cosmwasm_std::Env,
-                    _msg: cosmwasm_std::Reply,
-                ) -> anyhow::Result<cosmwasm_std::Response<cosmwasm_std::Empty>> {
-                    anyhow::bail!("reply not implemented for contract")
+                    _deps: #sylvia ::cw_std::DepsMut<#sylvia ::cw_std::Empty>,
+                    _env: #sylvia ::cw_std::Env,
+                    _msg: #sylvia ::cw_std::Reply,
+                ) -> #sylvia ::anyhow::Result<#sylvia ::cw_std::Response<#sylvia ::cw_std::Empty>> {
+                    #sylvia ::anyhow::bail!("reply not implemented for contract")
                 }
 
                 fn migrate(
                     &self,
-                    deps: cosmwasm_std::DepsMut<cosmwasm_std::Empty>,
-                    env: cosmwasm_std::Env,
+                    deps: #sylvia ::cw_std::DepsMut<#sylvia ::cw_std::Empty>,
+                    env: #sylvia ::cw_std::Env,
                     msg: Vec<u8>,
-                ) -> anyhow::Result<cosmwasm_std::Response<cosmwasm_std::Empty>> {
+                ) -> #sylvia ::anyhow::Result<#sylvia ::cw_std::Response<#sylvia ::cw_std::Empty>> {
                     #migrate_body
+                }
+            }
+        }
+    }
+}
+
+pub struct TraitMultitestHelpers<'a> {
+    trait_name: &'a Ident,
+}
+
+impl<'a> TraitMultitestHelpers<'a> {
+    pub fn new(source: &'a ItemTrait) -> Self {
+        Self {
+            trait_name: &source.ident,
+        }
+    }
+
+    pub fn emit(&self) -> TokenStream {
+        let trait_name = self.trait_name;
+        let sylvia = crate_module();
+        let proxy_name = Ident::new(&format!("{}Proxy", trait_name), trait_name.span());
+
+        quote! {
+            pub mod trait_utils {
+                pub struct #proxy_name <'app> {
+                    pub contract_addr: #sylvia ::cw_std::Addr,
+                    pub app: &'app #sylvia ::multitest::App,
+                }
+                impl<'app> #proxy_name <'app> {
+                    pub fn new(contract_addr: #sylvia ::cw_std::Addr, app: &'app #sylvia ::multitest::App) -> Self {
+                        #proxy_name { contract_addr, app }
+                    }
+                }
+                impl Into<#sylvia ::cw_std::Addr> for #proxy_name <'_> {
+                    fn into(self) -> #sylvia ::cw_std::Addr {
+                        self.contract_addr
+                    }
                 }
             }
         }