+++ CHANGELOG.md
@@ -4,6 +4,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- Changed the way multitest helpers are generated to avoid weird `use` statements in code.
+
 ## \[0.3.1\] - 2023-03-03
 
 - Slight improvement the invalid message received error